refactor: 칼럼 타입 코드 삭제 및 enum클래스 폴더 이동
