+++ src/main/java/com/hri/hri_web_backend/repository/PublicationRepository.java
@@ -2,14 +2,11 @@ package com.hri.hri_web_backend.repository;
 
 import java.util.List;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.hri.hri_web_backend.domain.Publication;
-import com.hri.hri_web_backend.domain.PublicationType;
+import com.hri.hri_web_backend.global.PublicationType;
 
 @Repository
 public interface PublicationRepository extends JpaRepository<Publication, Long> {
