+++ src/main/java/com/hri/hri_web_backend/service/PublicationService.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.hri.hri_web_backend.domain.Publication;
-import com.hri.hri_web_backend.domain.PublicationType;
+import com.hri.hri_web_backend.global.PublicationType;
 import com.hri.hri_web_backend.dto.PublicationDto;
 import com.hri.hri_web_backend.repository.PublicationRepository;
